django/requirements.txt
Babel==2.7.0
boto3==1.9.218
botocore==1.12.218
Django==1.11.23
django-phonenumber-field==2.4.0
django-storages==1.7.1
djangorestframework==3.9.4
docutils==0.15.2
futures==3.3.0
jmespath==0.9.4
phonenumbers==8.10.17
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
